refactor(design-picker): move ReadyDesignPicker to shared folder, use helper
Ready page now calls resolveDesignModuleProps instead of inline track
logic (pure refactor — identical props passed to ReadyDesignPicker).
ReadyDesignPicker.tsx relocates from the ready route to
components/clients/design-picker/ so the claimed-dashboard surface (next
task) can import it too; content unchanged except the action import path.
No re-export shim — grep confirmed the ready page was the only importer.
